+++ History.md
@@ -1,3 +1,8 @@
+### 0.0.3 / 2014-03-25
+
+* add `filter` setting that allows selecting fields that should be copied
+* skip a source if the source is a falsy value
+
 ### 0.0.2 / 2013-09-06
 
 * add `Gruntfile`